Split renderer and renderer types
